requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.0
Flask-Migrate==4.0.5
mysql-connector-python==8.4.0
python-dotenv==1.0.1

Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1

pandas==2.2.2
openpyxl==3.1.2

Werkzeug>=3.1.0,<4.0

# For CI/testing -- optional, recommended for Jenkins jobs
pytest==8.2.2
pytest-cov==5.0.0
